# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "RAFT"
  },
  "host": "raftsride.com/api",
  "basePath": "",
  "schemes": ["https", "http"],
  "paths": {
...
    "/vehicletypes/list": {
      "get": {
        "tags": ["System Settings"],
        "summary": "Vehicle Types",
        "description": "",
        "operationId": "vehicletypes",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Top Navigation Details"
          },
          "400": {
            "description": "Some problems occurred, please try again"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/razorpaydetails/list": {
      "get": {
        "tags": ["System Settings"],
        "summary": "Razor Pay Details",
        "description": "",
        "operationId": "razorpaydetails",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Top Navigation Details"
          },
          "400": {
            "description": "Some problems occurred, please try again"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  }
}